+++ Changes.md
@@ -3,6 +3,7 @@ WIP  TBD
  * Added the `version` command to track which version is installed.
  * Added the `openscripture.today` (with `ost` alias) to allow for showing the scripture of the day from [openscripture.today](https://openscripture.today).
  * Added the API library in `pkg/ost` for working with the openscripture.today API.
+ * Fix: There was a bug were resolution did not properly check to ensure that the verse was in the canon.
 
 0.0.0  2024-01-17